@@ -18,46 +18,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.7",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.7",
-    "@angular/compiler": "^7.2.7",
-    "@angular/core": "^7.2.7",
-    "@angular/forms": "^7.2.7",
-    "@angular/platform-browser": "^7.2.7",
-    "@angular/platform-browser-dynamic": "^7.2.7",
-    "@angular/router": "^7.2.7",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "alife-rating-star": "0.0.1",
-    "aws-amplify": "^1.1.19",
-    "aws-amplify-angular": "^2.1.6",
-    "bulma": "^0.7.4",
+    "aws-amplify": "^1.1.28",
+    "aws-amplify-angular": "^3.0.3",
+    "bulma": "^0.7.5",
     "bulma-slider": "^2.0.0",
     "bulma-tagsinput": "^2.0.0",
-    "core-js": "^2.6.5",
+    "core-js": "^2.6.9",
     "font-awesome": "^4.7.0",
     "moment": "^2.24.0",
-    "rxjs": "^6.4.0",
-    "typescript": "^3.2.4",
-    "webpack": "^4.29.5",
-    "zone.js": "^0.8.29"
+    "rxjs": "^6.5.2",
+    "typescript": "^3.4.5",
+    "webpack": "^4.32.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.3",
-    "@angular/cli": "^7.3.3",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "^7.2.7",
-    "@types/jest": "^24.0.9",
-    "@types/mocha": "^5.2.6",
-    "@types/node": "^10.12.27",
-    "codelyzer": "^4.3.0",
+    "@angular-devkit/build-angular": "^0.800.1",
+    "@angular/cli": "^8.0.1",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
+    "@types/jest": "^24.0.13",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^12.0.4",
+    "codelyzer": "^5.0.1",
     "flatted": "^2.0.0",
-    "jasmine-core": "~3.3.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^24.1.0",
-    "jest-preset-angular": "^6.0.2",
-    "jsdom": "^13.2.0",
+    "jest": "^24.8.0",
+    "jest-preset-angular": "^7.1.1",
+    "jsdom": "^15.1.1",
     "protractor": "^5.4.2",
-    "ts-node": "~7.0.1",
-    "tslint": "~5.11.0"
+    "ts-node": "~8.2.0",
+    "tslint": "~5.17.0"
   }
 }
